Use pachyderm v1.2.2
This seems to fix a few bugs. Putfile wasn't working w 1.2.3, now it is.
Also I think the pull policy was wrong or something for the job image ...
I think I was getting a cached version of the dowager image I was building,
which means I may not have been getting the job-shim fixes. Now I no longer
get the PPS_POD_NAME missing error on job pod startup.

Instead, I get an error in the train job that is totally valid. I'm not
supplying quite the right files in the scripts repo. Easy enough to fix that